% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel-aff.R
\name{aff}
\alias{aff}
\title{Calculate relative affinities for a fit}
\usage{
aff(fits, index)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Calculate relative affinities for a fit
}
\examples{

}
