% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-traj.R
\name{animate.traj}
\alias{animate.traj}
\title{Animate solution trajectory with PSAMs}
\usage{
animate.traj(trajPath, outPath, k)
}
\value{

}
\description{
Animate solution trajectory with PSAMs
}
\examples{

}
