% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly-enhancers.R
\name{fly.enhancers}
\alias{fly.enhancers}
\title{Define the set fly enhancers discussed in NRLB paper}
\usage{
fly.enhancers()
}
\value{
Nested list containing DNA sequence and other information
}
\description{
Define the set fly enhancers discussed in NRLB paper
}
\examples{


}
