% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{Generate PSAM}
\usage{
logo(fits, index = NULL, mode = NULL, rc = FALSE, display = TRUE,
  betas = NULL, save.path = NULL, isPDF = TRUE, title = NULL,
  ylim = NULL, fit.name = NULL, l.pad = 0, r.pad = 0, l.del = 0,
  r.del = 0)
}
\arguments{
\item{fits}{}

\item{index}{}

\item{mode}{}

\item{rc}{}
}
\value{

}
\description{
Generate PSAM
}
\examples{

}
