% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-seq.R
\name{max.seq}
\alias{max.seq}
\title{Find the highest affinity sequence for a model}
\usage{
\method{max}{seq}(fits, index, mode = NULL)
}
\value{

}
\description{
Find the highest affinity sequence for a model
}
\note{
uses dynamic programming
}
\examples{

}
