\name{MultinomialMethods-package}
\alias{MultinomialMethods-package}
\alias{MultinomialMethods}
\docType{package}
\title{
\packageTitle{MultinomialMethods}
}
\description{
\packageDescription{MultinomialMethods}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MultinomialMethods}
\packageIndices{MultinomialMethods}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MultinomialMethods}

Maintainer: \packageMaintainer{MultinomialMethods}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
