% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-NRLB.R
\docType{package}
\name{NRLB}
\alias{NRLB}
\alias{NRLB-package}
\title{NRLB: utitilies for working with NRLB models}
\description{
The NRLB package provides tools for working with protein-DNA recognition models created by the No Read Left Behind (NRLB) algorithm developed by the Bussemaker Lab in the Department of Biological Sciences at Columbia University.
}
\section{Foo functions}{

The foo functions ...
}

