% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-fit-results.R
\name{o}
\alias{o}
\title{Order fit results}
\usage{
o(fits, n = 5)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{n}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Order fit results
}
\examples{

}
