% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot.compare}
\alias{plot.compare}
\title{Compare two models}
\usage{
plot.compare(fitA, fitB, indexA = NULL, indexB = NULL, modeIdxA = NULL,
  modeIdxB = NULL, rcA = FALSE, rcB = FALSE, offset = NULL,
  labA = NULL, labB = NULL)
}
\arguments{
\item{fitA}{TO_BE_ADDED}

\item{fitB}{TO_BE_ADDED}

\item{indexA}{TO_BE_ADDED}

\item{indexB}{TO_BE_ADDED}

\item{modeIdxA}{TO_BE_ADDED}

\item{modeIdxB}{TO_BE_ADDED}

\item{rcA}{TO_BE_ADDED}

\item{rcB}{TO_BE_ADDED}

\item{offset}{TO_BE_ADDED}

\item{labA}{TO_BE_ADDED}

\item{labB}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Compare two models
}
\examples{

}
