% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likelihoods.R
\name{plot.likelihoods}
\alias{plot.likelihoods}
\title{Plot likelihoods}
\usage{
plot.likelihoods(fits, l, test = FALSE)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{l}{TO_BE_ADDED}

\item{test}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Plot likelihoods
}
\note{
does not work with multi-mode fits (notion of window is unknown)
}
\examples{

}
