% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-R0-qq.R
\name{plot.R0.qq}
\alias{plot.R0.qq}
\title{R0 QQ Plot}
\usage{
\method{plot}{R0.qq}(fileName, lowerBound, upperBound, divSize,
  completeCoverage = TRUE, compare = NULL, ylim = NULL, xlim = NULL)
}
\value{

}
\description{
R0 QQ Plot
}
\examples{

}
