% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-genome.R
\name{plot.score.genome}
\alias{plot.score.genome}
\title{Plot affinities of a sequence}
\usage{
\method{plot}{score.genome}(genomicSequence, fits, index, mode = NULL,
  rc = FALSE, nPeaks = NULL, annotate = FALSE, rescale = NULL,
  genomicSequence2 = NULL)
}
\value{

}
\description{
Plot affinities of a sequence
}
\examples{

}
