% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-fit.R
\name{print.fit}
\alias{print.fit}
\title{Print fit results}
\usage{
\method{print}{fit}(fits, index = NULL)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Print fit results
}
\examples{

}
