% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-logo.R
\name{ql}
\alias{ql}
\title{Quick Logo}
\usage{
ql(fits, index, mode = NULL, rc = FALSE)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}

\item{mode}{TO_BE_ADDED}

\item{rc}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Quick Logo
}
\examples{

}
