% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-genome.R
\name{score.genome}
\alias{score.genome}
\title{Score genome with a fit}
\usage{
score.genome(genomicSequence, fits, index, mode = NULL, rc = FALSE)
}
\value{

}
\description{
Score genome with a fit
}
\examples{

}
