% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-parser.R
\name{score.parser}
\alias{score.parser}
\title{Read in binary files containing genomic scoring results}
\usage{
score.parser(genome, score.dir, prot.name, chr.name, k)
}
\arguments{
\item{genome}{}

\item{chr.name}{}

\item{k}{}
}
\value{

}
\description{
Read in binary files containing genomic scoring results
}
\examples{

}
