% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-seq.R
\name{score.seq}
\alias{score.seq}
\title{Score DNA sequence using a model}
\usage{
score.seq(sequence, rSequence, k, isDinuc, nuc, dinuc)
}
\description{
Score DNA sequence using a model
}
\note{
requires explicit nuc, dinuc formatting (mode independent)
}
\examples{

}
