% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate-traj.R
\name{animate.traj}
\alias{animate.traj}
\title{Animate solution trajectory with PSAMs}
\usage{
animate.traj(trajPath, outPath, k)
}
\arguments{
\item{trajPath}{TO_BE_ADDED}

\item{outPath}{TO_BE_ADDED}

\item{k}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Animate solution trajectory with PSAMs
}
\examples{

}
