% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot.compare}
\alias{plot.compare}
\title{Compare two models}
\usage{
\method{plot}{compare}(fitA, fitB, indexA = NULL, indexB = NULL,
  modeIdxA = NULL, modeIdxB = NULL, rcA = FALSE, rcB = FALSE,
  offset = NULL, labA = NULL, labB = NULL)
}
\arguments{
\item{fitA}{}

\item{fitB}{}

\item{indexA}{}

\item{indexB}{}

\item{modeIdxA}{}

\item{modeIdxB}{}

\item{rcA}{}

\item{rcB}{}
}
\value{

}
\description{
Compare two models
}
\examples{

}
