% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-likelihoods.R
\name{plot.likelihoods}
\alias{plot.likelihoods}
\title{Plot likelihoods}
\usage{
\method{plot}{likelihoods}(fits, l, test = FALSE)
}
\description{
Plot likelihoods
}
\note{
does not work with multi-mode fits (notion of window is unknown)
}
\examples{

}
