% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-R0-qq.R
\name{plot.R0.qq}
\alias{plot.R0.qq}
\title{R0 QQ Plot}
\usage{
plot.R0.qq(fileName, lowerBound, upperBound, divSize, completeCoverage = TRUE,
  compare = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{fileName}{TO_BE_ADDED}

\item{lowerBound}{TO_BE_ADDED}

\item{upperBound}{TO_BE_ADDED}

\item{divSize}{TO_BE_ADDED}

\item{completeCoverage}{TO_BE_ADDED}

\item{compare}{TO_BE_ADDED}

\item{ylim}{TO_BE_ADDED}

\item{xlim}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
R0 QQ Plot
}
\examples{

}
